fix(codegen): assert the slot-aligned precondition the shape-region slack proof rests on; tighten the two-alignment-constants rationale
Polish round bgarg-number-20260904-fix1 answering the hotfix code review
(0 blockers): shape_frame_region_ptr now debug-asserts the raw region
byte offset is a multiple of FRAME_LOCAL_SLOT_SIZE — the slack slot
covers a round-up of at most one slot only under that precondition,
which the layout documented but never checked; FRAME_SHAPE_REGION_ALIGN
is u32 like FRAME_I128_SLOT_ALIGN; the comment on why two constants
coexist now gives the real reasons (claim_i128_align_by_provenance
already downgrades escaped i128 frame pointers; rounding cannot replace
claim-8 because the return slot at ABI offset 16 and the staging slot
cannot move, and shape consumers are type-generic field loads shared
with heap callers); build_module's ABI-alignment refusal is marked
unreachable-today; the IMP paragraph drops its "resolves once this
commit lands" clause now that git log --grep=bgarg-number resolves.

Gate on 0f1fd8a: fmt, clippy --all-targets clean on codegen/runtime/
driver, hotfix 6/6, frame_layouts_query 9, codegen lib 14, runtime lib
105, integration 530, the three M6 frame/number suites 41/18/31,
cross_impl_consistency 7 with the three new fixtures, release build,
alloc=free 4/4 armed, gitleaks clean; the RED re-proven on a clean main
checkout at ec014d8 — both fixtures signal 11, nested clean.

+++ crates/ynz-codegen/src/emit.rs
@@ -1213,7 +1213,9 @@ fn build_module<'ctx, 'g>(
         // must raise the constant (and the slack) in the same change.
         for (name, &struct_ty) in &shape_types.named {
             let align = u64::from(target_data.get_abi_alignment(&struct_ty));
-            if align > state_machine::FRAME_SHAPE_REGION_ALIGN {
+            // Unreachable from any Yinz program today (the widest field type is i128); fires
+            // loudly the first time a wider field type is lowered — no test exercises it.
+            if align > u64::from(state_machine::FRAME_SHAPE_REGION_ALIGN) {
                 return Err(format!(
                     "codegen: shape `{name}` has LLVM ABI alignment {align}, wider than the \
                      {} bytes a frame-embedded shape region is rounded to \
